Move LazyListState extension to a separate file
