+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/nft/collection/activity/NftCollectionEventsFragment.kt
@@ -29,8 +29,8 @@ import io.horizontalsystems.bankwallet.entities.ViewState
 import io.horizontalsystems.bankwallet.modules.coin.overview.Loading
 import io.horizontalsystems.bankwallet.modules.nft.NftCollection
 import io.horizontalsystems.bankwallet.modules.nft.collection.NftCollectionViewModel
-import io.horizontalsystems.bankwallet.modules.nft.collection.assets.OnBottomReached
 import io.horizontalsystems.bankwallet.ui.compose.ComposeAppTheme
+import io.horizontalsystems.bankwallet.ui.compose.OnBottomReached
 import io.horizontalsystems.bankwallet.ui.compose.components.*
 import io.horizontalsystems.core.findNavController
 import io.horizontalsystems.core.helpers.DateHelper
