+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/nft/collection/assets/NftCollectionAssetsFragment.kt
@@ -7,9 +7,8 @@ import android.view.ViewGroup
 import androidx.compose.animation.Crossfade
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.LazyListState
 import androidx.compose.foundation.lazy.rememberLazyListState
-import androidx.compose.runtime.*
+import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.ComposeView
 import androidx.compose.ui.platform.ViewCompositionStrategy
@@ -28,6 +27,7 @@ import io.horizontalsystems.bankwallet.modules.nft.asset.NftAssetModule
 import io.horizontalsystems.bankwallet.modules.nft.collection.NftCollectionViewModel
 import io.horizontalsystems.bankwallet.modules.nft.ui.NftAssetPreview
 import io.horizontalsystems.bankwallet.ui.compose.ComposeAppTheme
+import io.horizontalsystems.bankwallet.ui.compose.OnBottomReached
 import io.horizontalsystems.bankwallet.ui.compose.components.HSCircularProgressIndicator
 import io.horizontalsystems.bankwallet.ui.compose.components.ListErrorView
 import io.horizontalsystems.core.findNavController
@@ -136,32 +136,5 @@ private fun NftAssets(
 
     listState.OnBottomReached(buffer = 6) {
         onBottomReached()
-    }
-}
-
-@Composable
-fun LazyListState.OnBottomReached(
-    // tells how many items before we reach the bottom of the list
-    // to call onLoadMore function
-    buffer: Int = 0,
-    onLoadMore: () -> Unit
-) {
-    // Buffer must be positive.
-    // Or our list will never reach the bottom.
-    require(buffer >= 0) { "buffer cannot be negative, but was $buffer" }
-
-    val shouldLoadMore = remember {
-        derivedStateOf {
-            val lastVisibleItem = layoutInfo.visibleItemsInfo.lastOrNull()
-                ?: return@derivedStateOf true
-
-            // subtract buffer from the total items
-            lastVisibleItem.index >= layoutInfo.totalItemsCount - 1 - buffer
-        }
-    }
-
-    LaunchedEffect(shouldLoadMore) {
-        snapshotFlow { shouldLoadMore.value }
-            .collect { if (it) onLoadMore() }
     }
 }
