+++ app/src/main/java/io/horizontalsystems/bankwallet/ui/compose/LazyListState.kt
@@ -1,0 +1,31 @@
+package io.horizontalsystems.bankwallet.ui.compose
+
+import androidx.compose.foundation.lazy.LazyListState
+import androidx.compose.runtime.*
+
+@Composable
+fun LazyListState.OnBottomReached(
+    // tells how many items before we reach the bottom of the list
+    // to call onLoadMore function
+    buffer: Int = 0,
+    onLoadMore: () -> Unit
+) {
+    // Buffer must be positive.
+    // Or our list will never reach the bottom.
+    require(buffer >= 0) { "buffer cannot be negative, but was $buffer" }
+
+    val shouldLoadMore = remember {
+        derivedStateOf {
+            val lastVisibleItem = layoutInfo.visibleItemsInfo.lastOrNull()
+                ?: return@derivedStateOf true
+
+            // subtract buffer from the total items
+            lastVisibleItem.index >= layoutInfo.totalItemsCount - 1 - buffer
+        }
+    }
+
+    LaunchedEffect(shouldLoadMore) {
+        snapshotFlow { shouldLoadMore.value }
+            .collect { if (it) onLoadMore() }
+    }
+}
